# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E_Messe API",
        "description": "API pour l'application E_Messe - Gestion des messes",
        "contact": {
            "email": "leprodev03@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8081/api",
            "description": "Serveur Local"
        },
        {
            "url": "https://votre-domaine.com/api",
            "description": "Serveur de Production"
        }
    ],
    "paths": {
...
        "/api/events": {
            "get": {
                "tags": [
                    "Événements"
                ],
                "summary": "Récupérer la liste des événements disponibles pour l'utilisateur connecté",
                "description": "Retourne tous les événements actifs avec leurs détails principaux et la paroisse qui a créé l'événement.",
                "operationId": "getAllEvents",
                "responses": {
                    "200": {
                        "description": "Liste des événements récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "titre": {
                                                "type": "string",
                                                "example": "Concert de louange"
                                            },
                                            "date_debut": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-12-05T18:00:00"
                                            },
                                            "date_fin": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-12-05T21:00:00"
                                            },
                                            "lieu": {
                                                "type": "string",
                                                "example": "Cathédrale Saint-Paul"
                                            },
                                            "celebrant": {
                                                "type": "string",
                                                "example": "Père Jean Kouassi"
                                            },
                                            "participation_frais": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 2000
                                            },
                                            "image_url": {
                                                "type": "string",
                                                "example": "https://tonsite.com/storage/events_images/photo1.jpg"
                                            },
                                            "paroisse": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nom": {
                                                        "type": "string",
                                                        "example": "Paroisse Saint-Paul"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "paroisse@exemple.com"
                                                    },
                                                    "telephone": {
                                                        "type": "string",
                                                        "example": "+2250700000000"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Utilisateur non authentifié"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}